samples: bluetooth: Make hci_usb sample work with blueZ
BlueZ 5.50 (on ubuntu 19.10) assumes that extended advertising/scanning
commands are supported by default, when configuring the device after usb
enumeration.

This commit enables advertising extensions support for the nrf52840dk so
the sample will work without hiccups on a recent ubuntu linux system.

It also disables the VS command to switch to H4 transport because it is
not yet implemented in the Softdevice Controller.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>