drivers: uart: nrfx: Attempt receiving more than 1 byte
To not bother the CPU too much, I'd like to try receiving the number of
bytes requested through uart_fifo_read() in one DMA transaction.

The rx buffer size is set to the configured tx buffer size in order to
not introduce yet another Kconfig option that will have to be duplicated
for each peripheral instance.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>