Bluetooth: host: l2cap: give enough credits to receive an SDU
See exhaustive comment in the code.

The previous logic doesn't work if the remote uses more than 2 *
`init_credits` PDUs to transfer its SDU.

It still made the same assumptions about buffer space, but not explicitly
stated. In addition, the value to put inside `initial_credits` was not
documented.

This patch:
- makes it clear what we expect of `initial_credits`
- makes the host disconnect the channel if that condition is not fulfilled
- fixes `l2cap_chan_update_credits`'s credit calculation
- exempts the extra credits sent in that case from the normal checks

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>