Bluetooth: host: fix deadlock when `CONFIG_BT_RECV_WORKQ_BT=y`
A deadlock can occur in that situation:
- bt wq handles an HCI buffer that is an ATT REQ
- handling it involves sending a RSP
- sending a RSP involves waiting for a conn TX context
- bt wq waits forever
- the HCI buffer is freed _after_ the l2cap .chan cb
- in that case, that is att_recv, which is currently stuck
-> buffer is held up

- hci thread waits on a data buffer
-> last buffer is held by bt wq

The bug doesn't manifest itself when using the syswq for bt_recv.
This is infortunate as using the BT WQ is the default option.

To work around it, apply the same rule as the syswq: don't wait for a TX
context or ATT metadata when sending from the workqueue that bt_recv is
called from.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>