Bluetooth: l2cap: remove metadata allocation
The previous commit removed the use of `bt_l2cap_chan_send_cb`. That means that
the `cb` and `userdata` will always be NULL and thus can be omitted.

For the non-basic channels (use of `bt_l2cap_chan_send`), we changed
`bt_l2cap_send_cb()` so that it's called with userdata==CID instead of
allocating a `struct l2cap_tx_metadata` and passing in a pointer to that.

This allows us to delete the whole l2cap metadata storage and allocation code,
saving us some memory, code and avoiding an unnecessary allocation bottleneck.