trace host/att/eatt_notif
This one doesn't drop events but is a little bland.
edit: it does drop events in the perfetto UI.

But I figured out why tracy is getting confused:
Its stacktrace view gets confused when it gets a B+E event with a
different ID (on the same thread). It's really wired to think that
functions have to have an E event before "going" into another one.

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>