fix the host 🤠💣
@alwa:

Mental gymnastics are hard:

We were marking the conn->rx buf as NCP-sent. We shouldn't have, as
conn->rx is the buffer that is finally unref'd and makes flow control
tick.

The logic for sending the NCP is correct though: always send it for
partial bufs, except the last one.

In order to delay the last sent NCP, we should not mark conn->rx as
NCP-sent: that ensures the destroy() callback will send the final NCP,
restoring the flow.

Now your test fails properly, enjoy 🥳

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>