DX: make the first devcontainer
This PR adds 3 devcontainer [1] environments.
1. "Zephyr basic": run twister, check_compliance.py
2. "Bluetooth w/ babblesim" contribute to the Bluetooth subsystem.

The containers are based on the Zephyr CI image.

These environments are directly usable on Github codespaces [3], and
enable contributors to develop features, bug fixes and tests armed with
only a browser (and some $$$).

They are also usable locally, with Visual Studio Code. Tested on linux
with native docker, but should also work on Windows with WSL.

For this first version there are a few rough edges, notably:

- CI image is bloated with unnecessary toolchains
- Image has repo cache. Maybe faster if we just clone from scratch
- The ubuntu system in the image is "minimized". That means a bunch of QoL
features are disabled, like shell completion.

[1] What are devcontainers:
https://code.visualstudio.com/docs/devcontainers/create-dev-container

[2] Devcontainer specification:
https://containers.dev/implementors/json_reference/
https://containers.dev/implementors/spec/

[3] Codespaces (hosted devcontainer platform)
https://github.com/codespaces

Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>